refactor(testfiles): rename testdata files
