+++ src/utils/searchUtils.test.tsx
@@ -4,10 +4,10 @@ import type { FiltersWithArrays } from "./searchUtils";
 import { ABSENT_FILTER_TOKEN } from "./absent";
 import { Scenario, SearchFilters } from "../types";
 
-import sample01 from "../../testdata/valid/scenarios_metadata_sample_01.json" assert { type: "json" };
-import sample02 from "../../testdata/valid/scenarios_metadata_sample_02.json" assert { type: "json" };
-import sample03 from "../../testdata/valid/scenarios_metadata_sample_03.json" assert { type: "json" };
-import sample04 from "../../testdata/valid/scenarios_metadata_sample_04.json" assert { type: "json" };
+import sample01 from "../../testdata/valid/pathwayMetadata_sample_01.json" assert { type: "json" };
+import sample02 from "../../testdata/valid/pathwayMetadata_sample_02.json" assert { type: "json" };
+import sample03 from "../../testdata/valid/pathwayMetadata_sample_03.json" assert { type: "json" };
+import sample04 from "../../testdata/valid/pathwayMetadata_sample_04.json" assert { type: "json" };
 const mockScenarios: Scenario[] = [sample01, sample02, sample03, sample04];
 
 describe("searchUtils", () => {
@@ -148,7 +148,7 @@ describe("searchUtils", () => {
   });
 });
 
-import rawFullScenario from "../../testdata/valid/scenarios_metadata_full.json" assert { type: "json" };
+import rawFullScenario from "../../testdata/valid/pathwayMetadata_full.json" assert { type: "json" };
 const mockFullScenarios: Scenario[] = [rawFullScenario];
 
 describe("searchUtils - array results", () => {
