+++ testdata/valid/pathwayMetadata_full.json
