+++ testdata/valid/pathwayMetadata_minimal.json
