+++ testdata/valid/pathwayMetadata_sample_02.json
