+++ testdata/valid/pathwayMetadata_sample_03.json
