+++ testdata/valid/pathwayMetadata_sample_04.json
