+++ testdata/valid/pathwayMetadata_standard.json
